+++ ECommereceApi/Repo/UserRepo.cs
@@ -27,7 +27,7 @@ namespace ECommereceApi.Repo
 
         public async Task<IEnumerable<UserDTO>> GetAdminsAsync()
         {
-            var users = await _context.Users.Where(x=>x.Role==RoleType.Admin).ToListAsync();
+            var users = await _context.Users.Where(x => x.Role == RoleType.Admin).ToListAsync();
             return _mapper.Map<List<UserDTO>>(users);
         }
         public async Task<IEnumerable<UserDTO>> GetCustomersAsync()
@@ -45,17 +45,17 @@ namespace ECommereceApi.Repo
         public async Task<UserDTO?> GetSuperAdminAsync()
         {
             var super = _context.Users.First(x => x.Role == RoleType.Admin);
-            if(super == null)
+            if (super == null)
                 return null;
             return _mapper.Map<UserDTO>(super);
         }
 
         private async Task<bool> IsSuperAdmin(int id)
         {
-            var user = _context.Users.FirstOrDefault(x=>x.UserId==id);
-            if(user == null) return false;
+            var user = _context.Users.FirstOrDefault(x => x.UserId == id);
+            if (user == null) return false;
 
-            if(user.UserId==GetSuperAdminAsync().Id)
+            if (user.UserId == GetSuperAdminAsync().Id)
                 return true;
             return false;
 
@@ -72,7 +72,7 @@ namespace ECommereceApi.Repo
             var user = _mapper.Map<User>(userDto);
             user.IsVerified = true;
             user.VertificationCode = "adminver"; //  Generate Random Code
-          
+
             if (userDto.Role == RoleType.Customer)
             {
                 await _context.Users.AddAsync(user);
@@ -81,7 +81,7 @@ namespace ECommereceApi.Repo
                 {
                     return status;
                 }
-                await _context.Customers.AddAsync(new Customer { UserId = user.UserId});
+                await _context.Customers.AddAsync(new Customer { UserId = user.UserId });
                 return await SaveAsync();
             }
             else
@@ -97,12 +97,33 @@ namespace ECommereceApi.Repo
             }
         }
 
-        public async Task<Status> UpdateUserAsync(UserDTO userDto)
+        public async Task<Status> UpdateUserAsync(int userId, UserDTOUi userDtoUi)
         {
-            userDto.Email = userDto.Email.ToLower().Trim();
-            var user = _mapper.Map<User>(userDto);
-            _context.Entry(user).State = EntityState.Modified;
-            return await SaveAsync();
+            var user = await _context.Users.FindAsync(userId);
+            if (user == null)
+            {
+                return Status.NotFound;
+            }
+            if (await EmailExistedBeforeExceptCurrent(userDtoUi.Email, userId))
+            {
+                return Status.EmailExistsBefore;
+            }
+
+            _context.Update(user);
+            try
+            {
+                await _context.SaveChangesAsync();
+                return Status.Success;
+            }
+            catch
+            {
+                return Status.Failed;
+            }
+        }
+
+        private Task<bool> EmailExistedBeforeExceptCurrent(string email, int id)
+        {
+            return _context.Users.AnyAsync(u => u.Email == email && u.UserId != id);
         }
 
         public async Task<Status> DeleteUserAsync(int id)
@@ -111,6 +132,10 @@ namespace ECommereceApi.Repo
             if (user == null)
             {
                 return Status.NotFound;
+            }
+            if (IsSuperAdmin(user.UserId).Result)
+            {
+                return Status.SuperAdminConstraint;
             }
             _context.Users.Remove(user);
             return await SaveAsync();
@@ -184,7 +209,7 @@ namespace ECommereceApi.Repo
             // checking if the user is superAdmin you can't delete him
             if (IsSuperAdmin(user.UserId).Result)
             {
-                return Status.Failed;
+                return Status.SuperAdminConstraint;
             }
             user.IsDeleted = true;
             return await SaveAsync();
